Removing useFilters.ts from filter component directory, copied useFilters.ts to useURLPrarms.ts
